chore: move transform to pkg
